added tests that ensure unauthenticated requests are rejected when appropriate fixed security in ensuring category removed obsolete rls
